Split scene and sidebar support by responsibility
Oversized app files make ownership and safe review harder even when
their responsibilities are already separable.

Move existing coordinators, session state, lifecycle helpers, and
sidebar presentation models into files named for those responsibilities.
Keep stateful scene orchestration in place until it has a real collaborator
boundary instead of widening private access for an arbitrary split.
