+++ Sources/App/WorkspaceSceneModel.swift
@@ -41,469 +41,6 @@ private func presentGhosthubAlert(
 }
 #endif
 
-enum TmuxSessionThemeError: Error, Equatable, LocalizedError {
-    case unavailable(session: String)
-
-    var errorDescription: String? {
-        switch self {
-        case let .unavailable(session):
-            return "The theme cannot be applied to session “\(session)”"
-                + " because its active tmux attachment is unavailable."
-        }
-    }
-}
-
-enum HerdrSessionPresentationError: Error, Equatable, LocalizedError {
-    case unavailable
-    case sessionExists(String)
-    case sessionMissing(String)
-    case sessionNotRunning(String)
-    case sessionNotStopped(String)
-    case operationPending(String)
-    case routeChangedDuringValidation(String)
-    case stateChangedDuringValidation(String)
-    case stateValidationFailed(name: String, detail: String)
-
-    var errorDescription: String? {
-        switch self {
-        case .unavailable:
-            "Herdr is unavailable on this host."
-        case let .sessionExists(name):
-            "A Herdr session named “\(name)” already exists."
-        case let .sessionMissing(name):
-            "Herdr session “\(name)” no longer exists."
-        case let .sessionNotRunning(name):
-            "Herdr session “\(name)” is not running."
-        case let .sessionNotStopped(name):
-            "Herdr session “\(name)” is not stopped."
-        case let .operationPending(name):
-            "Another operation is already changing “\(name)”."
-        case let .routeChangedDuringValidation(name):
-            "The connection for the host containing “\(name)” changed while Ghosthub was checking it."
-        case let .stateChangedDuringValidation(name):
-            "Herdr session “\(name)” changed while Ghosthub was checking it. Try again."
-        case let .stateValidationFailed(name, detail):
-            "Ghosthub could not confirm the current state of Herdr session “\(name)”: \(detail)"
-        }
-    }
-}
-
-enum HerdrSessionLifecycleRequestError: Error, Equatable, LocalizedError {
-    case hostChanged(String)
-
-    var errorDescription: String? {
-        switch self {
-        case let .hostChanged(name):
-            "The connection for the host containing “\(name)” changed."
-        }
-    }
-}
-
-enum ZellijSessionPresentationError: Error, Equatable, LocalizedError {
-    case unavailable
-    case sessionExists(String)
-    case sessionMissing(String)
-    case hostChanged(String)
-    case operationPending(String)
-
-    var errorDescription: String? {
-        switch self {
-        case .unavailable:
-            "Zellij is unavailable on this host."
-        case let .sessionExists(name):
-            "A Zellij session named “\(name)” already exists."
-        case let .sessionMissing(name):
-            "Zellij session “\(name)” no longer exists."
-        case let .hostChanged(name):
-            "The connection for the host containing “\(name)” changed."
-        case let .operationPending(name):
-            "Another operation is already changing “\(name)”."
-        }
-    }
-}
-
-struct WorkspaceInventoryRefreshProgress: Equatable {
-    var kwtCompleted = false
-    var tmuxCompleted = false
-    var herdrCompleted = false
-    var zellijCompleted = false
-}
-
-enum NativeSessionRecoveryState: Equatable {
-    case reconnecting(message: String)
-    case needsAttention(message: String, canReviewConnection: Bool)
-
-    var isReconnecting: Bool {
-        if case .reconnecting = self {
-            return true
-        }
-        return false
-    }
-
-    var allowsReconnectNow: Bool {
-        switch self {
-        case .reconnecting:
-            true
-        case let .needsAttention(_, canReviewConnection):
-            !canReviewConnection
-        }
-    }
-}
-
-@MainActor
-final class WorktreeMutationCoordinator {
-    struct Scope: Hashable, Sendable {
-        let hostID: UUID
-        let projectIdentity: String
-    }
-
-    struct ProjectRegistryHost: Hashable, Sendable {
-        let target: CommandHost
-    }
-
-    struct ProtectedEndpoint: Hashable, Sendable {
-        let worktreeName: String
-        let worktreeIdentity: KwtWorktreeIdentity
-        let selection: WorkspaceTmuxSessionSelection?
-    }
-
-    typealias RemovalTombstone = KwtWorktreeIdentity
-
-    enum Phase: Sendable {
-        case began
-        case willRemove
-        case quarantined
-        case ended
-    }
-
-    struct Event: Sendable {
-        let phase: Phase
-        let scope: Scope
-        let removalTombstones: Set<RemovalTombstone>
-        let removalPresentationTargets: Set<WorkspaceTmuxSessionSelection>
-        /// Endpoints the failed removal's recovery refresh confirmed for
-        /// the surviving target; `nil` when no authoritative refresh
-        /// reconciled the failure.
-        let reconciledRestorationTargets: Set<WorkspaceTmuxSessionSelection>?
-        let requiresWorkspaceReestablishment: Bool
-        let removesProject: Bool
-        let allowsRemovalRestoration: Bool
-    }
-
-    struct QuarantinedProjectRemoval: Equatable, Sendable {
-        let projectPath: String
-        let host: CommandHost
-    }
-
-    static let shared = WorktreeMutationCoordinator()
-
-    private var activeScopes: Set<Scope> = []
-    private var activeProjectRegistryHosts: Set<ProjectRegistryHost> = []
-    private var projectRemovalRegistryHostsByScope:
-        [Scope: ProjectRegistryHost] = [:]
-    private var pendingRemovalsByScope: [Scope: Set<RemovalTombstone>] = [:]
-    private var quarantinedProjectRemovalsByScope:
-        [Scope: QuarantinedProjectRemoval] = [:]
-    private var protectedEndpointsByParticipant:
-        [UUID: [Scope: Set<ProtectedEndpoint>]] = [:]
-    private var retiredProtectedEndpointsByScope:
-        [Scope: Set<ProtectedEndpoint>] = [:]
-    private let eventSubject = PassthroughSubject<Event, Never>()
-
-    private struct ProtectedSelectionKey: Hashable {
-        let hostID: UUID
-        let name: String
-        let socketName: String?
-
-        init(_ selection: WorkspaceTmuxSessionSelection) {
-            hostID = selection.hostID
-            name = selection.name
-            socketName = selection.socketName
-        }
-    }
-
-    var events: AnyPublisher<Event, Never> {
-        eventSubject.eraseToAnyPublisher()
-    }
-
-    var scopes: Set<Scope> {
-        activeScopes
-    }
-
-    var pendingRemovals: [Scope: Set<RemovalTombstone>] {
-        pendingRemovalsByScope
-    }
-
-    var quarantinedProjectRemovals: [Scope: QuarantinedProjectRemoval] {
-        quarantinedProjectRemovalsByScope
-    }
-
-    func replaceProtectedEndpoints(
-        _ endpoints: [Scope: Set<ProtectedEndpoint>],
-        for participantID: UUID
-    ) {
-        for (scope, previous) in
-            protectedEndpointsByParticipant[participantID] ?? [:] {
-            let superseded = previous.subtracting(endpoints[scope] ?? [])
-            retiredProtectedEndpointsByScope[scope, default: []]
-                .formUnion(superseded)
-        }
-        for (scope, replacements) in endpoints {
-            let resolvedIdentities = Set(replacements.compactMap { endpoint in
-                endpoint.selection == nil ? nil : endpoint.worktreeIdentity
-            })
-            let resolvedSelections = Set(replacements.compactMap {
-                $0.selection.map(ProtectedSelectionKey.init)
-            })
-            guard let retired = retiredProtectedEndpointsByScope[scope]
-            else { continue }
-            let remaining = retired.filter { endpoint in
-                guard let selection = endpoint.selection else {
-                    return !resolvedIdentities.contains(
-                        endpoint.worktreeIdentity
-                    )
-                }
-                return !resolvedSelections.contains(
-                    ProtectedSelectionKey(selection)
-                )
-            }
-            if remaining.isEmpty {
-                retiredProtectedEndpointsByScope.removeValue(forKey: scope)
-            } else {
-                retiredProtectedEndpointsByScope[scope] = Set(remaining)
-            }
-        }
-        protectedEndpointsByParticipant[participantID] = endpoints
-    }
-
-    func retireProtectedEndpoints(for participantID: UUID) {
-        guard let endpoints = protectedEndpointsByParticipant.removeValue(
-            forKey: participantID
-        ) else { return }
-        for (scope, values) in endpoints {
-            retiredProtectedEndpointsByScope[scope, default: []]
-                .formUnion(values)
-        }
-    }
-
-    func confirmProtectedEndpointAbsent(
-        _ selection: WorkspaceTmuxSessionSelection,
-        in scope: Scope
-    ) {
-        guard let endpoints = retiredProtectedEndpointsByScope[scope]
-        else { return }
-        let remaining = Set(endpoints.filter { endpoint in
-            endpoint.selection != selection
-        })
-        if remaining.isEmpty {
-            retiredProtectedEndpointsByScope.removeValue(forKey: scope)
-        } else {
-            retiredProtectedEndpointsByScope[scope] = remaining
-        }
-    }
-
-    func reconcileRetiredProtectedEndpoints(
-        after inventory: KwtHostInventory,
-        hostID: UUID
-    ) {
-        guard inventory.projectsWarning == nil else { return }
-        let scopes = retiredProtectedEndpointsByScope.keys.filter {
-            $0.hostID == hostID
-        }
-        for scope in scopes {
-            guard let project = inventory.projects.first(where: {
-                $0.project.repository == scope.projectIdentity
-            }) else {
-                retiredProtectedEndpointsByScope.removeValue(forKey: scope)
-                continue
-            }
-            guard project.warning == nil,
-                  let retired = retiredProtectedEndpointsByScope[scope]
-            else { continue }
-            let remaining = retired.filter { endpoint in
-                project.worktrees.contains { worktree in
-                    endpoint.worktreeIdentity.matches(
-                        path: worktree.path,
-                        generation: worktree.generation
-                    )
-                }
-            }
-            if remaining.isEmpty {
-                retiredProtectedEndpointsByScope.removeValue(forKey: scope)
-            } else {
-                retiredProtectedEndpointsByScope[scope] = Set(remaining)
-            }
-        }
-    }
-
-    func protectedEndpoints(in scope: Scope) -> Set<ProtectedEndpoint> {
-        protectedEndpointsByParticipant.values.reduce(
-            into: retiredProtectedEndpointsByScope[scope] ?? []
-        ) {
-            $0.formUnion($1[scope] ?? [])
-        }
-    }
-
-    func acquire(
-        hostID: UUID,
-        projectIdentity: String
-    ) -> Bool {
-        let scope = Scope(
-            hostID: hostID,
-            projectIdentity: projectIdentity
-        )
-        let inserted = activeScopes.insert(scope).inserted
-        if inserted {
-            eventSubject.send(
-                Event(
-                    phase: .began,
-                    scope: scope,
-                    removalTombstones: [],
-                    removalPresentationTargets: [],
-                    reconciledRestorationTargets: nil,
-                    requiresWorkspaceReestablishment: false,
-                    removesProject: false,
-                    allowsRemovalRestoration: true
-                )
-            )
-        }
-        return inserted
-    }
-
-    func acquireProjectRegistry(host: ProjectRegistryHost) -> Bool {
-        activeProjectRegistryHosts.insert(host).inserted
-    }
-
-    func releaseProjectRegistry(host: ProjectRegistryHost) {
-        activeProjectRegistryHosts.remove(host)
-    }
-
-    func acquireProjectRemoval(
-        hostID: UUID,
-        projectIdentity: String,
-        registryHost: ProjectRegistryHost
-    ) -> Bool {
-        let scope = Scope(
-            hostID: hostID,
-            projectIdentity: projectIdentity
-        )
-        guard !activeProjectRegistryHosts.contains(registryHost),
-              !activeScopes.contains(scope)
-        else { return false }
-        activeProjectRegistryHosts.insert(registryHost)
-        projectRemovalRegistryHostsByScope[scope] = registryHost
-        activeScopes.insert(scope)
-        eventSubject.send(
-            Event(
-                phase: .began,
-                scope: scope,
-                removalTombstones: [],
-                removalPresentationTargets: [],
-                reconciledRestorationTargets: nil,
-                requiresWorkspaceReestablishment: false,
-                removesProject: false,
-                allowsRemovalRestoration: true
-            )
-        )
-        return true
-    }
-
-    func release(
-        hostID: UUID,
-        projectIdentity: String,
-        removalTombstones: Set<RemovalTombstone> = [],
-        reconciledRestorationTargets:
-        Set<WorkspaceTmuxSessionSelection>? = nil,
-        requiresWorkspaceReestablishment: Bool = false,
-        removesProject: Bool = false,
-        allowsRemovalRestoration: Bool = true
-    ) {
-        let scope = Scope(
-            hostID: hostID,
-            projectIdentity: projectIdentity
-        )
-        if activeScopes.remove(scope) != nil {
-            if let registryHost = projectRemovalRegistryHostsByScope
-                .removeValue(forKey: scope) {
-                activeProjectRegistryHosts.remove(registryHost)
-            }
-            pendingRemovalsByScope.removeValue(forKey: scope)
-            quarantinedProjectRemovalsByScope.removeValue(forKey: scope)
-            eventSubject.send(
-                Event(
-                    phase: .ended,
-                    scope: scope,
-                    removalTombstones: removalTombstones,
-                    removalPresentationTargets: [],
-                    reconciledRestorationTargets:
-                    reconciledRestorationTargets,
-                    requiresWorkspaceReestablishment:
-                    requiresWorkspaceReestablishment,
-                    removesProject: removesProject,
-                    allowsRemovalRestoration: allowsRemovalRestoration
-                )
-            )
-        }
-    }
-
-    func prepareRemoval(
-        hostID: UUID,
-        projectIdentity: String,
-        worktrees: Set<RemovalTombstone>,
-        presentationTargets: Set<WorkspaceTmuxSessionSelection>
-    ) {
-        let scope = Scope(
-            hostID: hostID,
-            projectIdentity: projectIdentity
-        )
-        guard activeScopes.contains(scope) else { return }
-        pendingRemovalsByScope[scope, default: []].formUnion(worktrees)
-        eventSubject.send(
-            Event(
-                phase: .willRemove,
-                scope: scope,
-                removalTombstones: worktrees,
-                removalPresentationTargets: presentationTargets,
-                reconciledRestorationTargets: nil,
-                requiresWorkspaceReestablishment: false,
-                removesProject: false,
-                allowsRemovalRestoration: true
-            )
-        )
-    }
-
-    func quarantineProjectRemoval(
-        hostID: UUID,
-        projectIdentity: String,
-        projectPath: String,
-        host: CommandHost
-    ) {
-        let scope = Scope(
-            hostID: hostID,
-            projectIdentity: projectIdentity
-        )
-        guard activeScopes.contains(scope),
-              pendingRemovalsByScope[scope] != nil
-        else { return }
-        quarantinedProjectRemovalsByScope[scope] = QuarantinedProjectRemoval(
-            projectPath: projectPath,
-            host: host
-        )
-        eventSubject.send(
-            Event(
-                phase: .quarantined,
-                scope: scope,
-                removalTombstones: [],
-                removalPresentationTargets: [],
-                reconciledRestorationTargets: nil,
-                requiresWorkspaceReestablishment: false,
-                removesProject: false,
-                allowsRemovalRestoration: false
-            )
-        )
-    }
-}
-
 @MainActor
 final class WorkspaceSceneModel: ObservableObject {
     typealias KwtInventoryLoader = @Sendable (
